insights: support for pie charts in the GraphQL API (#27311)

+++ internal/insights/insights.go
@@ -276,7 +276,7 @@ type LangStatsInsight struct {
 	ID             string
 	Title          string
 	Repository     string
-	OtherThreshold float32
+	OtherThreshold float64
 	OrgID          *int32
 	UserID         *int32
 }
